geotrade/backend/requirements.txt
# Core
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
python-dotenv==1.0.1

# Async DB clients
asyncpg==0.29.0
motor==3.4.0
redis[asyncio]==5.0.4

# Kafka
aiokafka==0.10.0

# NLP / ML
transformers==4.40.1
torch==2.3.0
spacy>=3.8.0
scikit-learn==1.4.2
xgboost==2.0.3
statsmodels==0.14.2
numpy==1.26.4
pandas==2.2.2

# Crawling
scrapy==2.11.1
beautifulsoup4==4.12.3
feedparser==6.0.11
httpx==0.27.0

# AWS
boto3==1.34.84
botocore==1.34.84

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9